fix: Remove default API base URL
For some reason, the default api.hookdeck.com URL without versioning is running into an issue. I'm removing that from the provider and instead use the latest version from the SDK instead. If people wants to override this value, it's recommended that they include the versioning in their base URL as well.