feat: Update to Hookdeck Go SDK v0.0.37 and support more source verifications (#28)
* chore: Update Hookdeck Go SDK to v0.0.37

* fix: Rename "archived" to "disabled"

* fix: Update source verification change in SDK

* fix: Update source verification config naming convention

* fix: Remove default API base URL

For some reason, the default api.hookdeck.com URL without versioning is running into an issue. I'm removing that from the provider and instead use the latest version from the SDK instead. If people wants to override this value, it's recommended that they include the versioning in their base URL as well.

* chore: Add optional to disabled_at field

* fix: Include type to Verification payload

* fix: Recharge source verification

* feat: Add new source verifications

* docs: Generate docs

* feat: Add JSON source verification

* docs: Generate docs

* lint: Add error handling around JSON unmarshalling